+++ groups/src/jvmMain/kotlin/inkapplications/shade/groups/KtorHueGroupsApi.kt
@@ -1,11 +1,11 @@
 package inkapplications.shade.groups
 
-import inkapplications.shade.constructs.HueProperties
 import inkapplications.shade.constructs.HueResponse
 import inkapplications.shade.constructs.IdToken
 import io.ktor.client.*
 import io.ktor.client.request.*
 import io.ktor.http.*
+import kotlinx.serialization.json.JsonObject
 import kotlin.time.ExperimentalTime
 
 @ExperimentalTime
@@ -32,7 +32,7 @@ internal class KtorHueGroupsApi(
         }
     }
 
-    override suspend fun updateGroup(token: String, groupId: String, attributes: MutableGroupAttributes): HueResponse<HueProperties> {
+    override suspend fun updateGroup(token: String, groupId: String, attributes: MutableGroupAttributes): HueResponse<JsonObject> {
         return client.put("api/$token/groups/$groupId") {
             accept(ContentType.Application.Json)
             contentType(ContentType.Application.Json)
@@ -40,7 +40,7 @@ internal class KtorHueGroupsApi(
         }
     }
 
-    override suspend fun setState(token: String, groupId: String, state: GroupStateModification): HueResponse<HueProperties> {
+    override suspend fun setState(token: String, groupId: String, state: GroupStateModification): HueResponse<JsonObject> {
         return client.put(createHueGroupsStateUrl(token, groupId)) {
             accept(ContentType.Application.Json)
             contentType(ContentType.Application.Json)